remove unnecessary pages
