--- src/app/[locale]/cafe/[id]/layout.tsx
@@ -1,19 +1,0 @@
-import { desktop, mobile } from "@/utils/const";
-import { Box } from "@mui/material";
-
-export default function CafeListLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <Box
-      sx={{
-        mt: { xs: mobile.space.aroundY, md: desktop.space.aroundY },
-        mb: 6,
-      }}
-    >
-      {children}
-    </Box>
-  );
-}
