--- src/data/areas.js
@@ -1,18 +1,0 @@
-export const areaInfo = [
-  {
-    title: "Nakameguro, Daikanyama, Ebisu Area",
-    stations: ["中目黒駅", "代官山駅", "恵比寿駅"],
-  },
-  {
-    title: "Kichijoji Area",
-    stations: ["吉祥寺駅", "三鷹駅"],
-  },
-  {
-    title: "Omotesando Area",
-    stations: ["外苑前駅", "表参道駅", "原宿駅", "青山一丁目駅"],
-  },
-  {
-    title: "Yoyogi-Uehara Area",
-    stations: ["代々木駅", "代々木上原駅", "代々木八幡駅"],
-  },
-];
