--- src/data/buttons.js
@@ -1,18 +1,0 @@
-export const loginProviders = [
-  {
-    text: "Google",
-    iconUrl: "/login/loginIcon/google.svg",
-    signInUrl: "google",
-  },
-  {
-    text: "FaceBook",
-    iconUrl: "/login/loginIcon/facebook.svg",
-    signInUrl: "facebook",
-  },
-  {
-    text: "GitHub",
-    iconUrl: "/login/loginIcon/github.svg",
-    signInUrl: "github",
-  },
-  { text: "LINE", iconUrl: "/login/loginIcon/line.svg", signInUrl: "line" },
-];
