--- src/data/cafeCheckKeyword.js
@@ -1,93 +1,0 @@
-export const cafeKeywords = [
-  "cafe",
-  "coffee",
-  "コーヒー",
-  "喫茶店",
-  "カフェ",
-  "espresso",
-  "latte",
-  "cappuccino",
-  "americano",
-  "macchiato",
-  "mocha",
-  "brew",
-  "barista",
-  "beans",
-  "roast",
-  "grind",
-  "drip",
-  "filter",
-  "frappe",
-  "decaf",
-  "cold brew",
-  "iced coffee",
-  "フレンチプレス",
-  "カプチーノ",
-  "ラテ",
-  "エスプレッソ",
-  "アメリカーノ",
-  "マキアート",
-  "モカ",
-  "ブレンド",
-  "バリスタ",
-  "ビーンズ",
-  "焙煎",
-  "グラインド",
-  "ドリップ",
-  "フィルター",
-  "フラペ",
-  "デカフェ",
-  "コールドブリュー",
-  "アイスコーヒー",
-  "テイクアウト",
-  "テーブル",
-  "カウンター",
-  "席",
-  "椅子",
-  "飲み物",
-  "フード",
-  "ペストリー",
-  "ケーキ",
-  "パイ",
-  "サンドイッチ",
-  "スイーツ",
-  "ブレックファスト",
-  "ランチ",
-  "ディナー",
-  "リラックス",
-  "ミーティング",
-  "読書",
-  "勉強",
-  "仕事",
-  "休憩",
-  "友達",
-  "デート",
-  "おしゃれ",
-  "インテリア",
-  "デザイン",
-  "雰囲気",
-  "アート",
-  "音楽",
-  "ジャズ",
-  "コーヒー豆",
-  "ハンドドリップ",
-  "カスタマイズ",
-  "メニュー",
-  "レシピ",
-  "トースト",
-  "マフィン",
-  "クロワッサン",
-  "デニッシュ",
-  "カップ",
-  "ソーサー",
-  "ミルク",
-  "シュガー",
-  "クリーム",
-  "スプーン",
-  "ナプキン",
-  "トレイ",
-  "ブランチ",
-  "午後のひととき",
-  "おもてなし",
-  "リラックスタイム",
-];
