--- src/data/stations.json
@@ -1,146 +1,0 @@
-[
-  { "ja": "東京駅", "en": "Tokyo Station" },
-  { "ja": "新橋駅", "en": "Shimbashi Station" },
-  { "ja": "品川駅", "en": "Shinagawa Station" },
-  { "ja": "代々木上原駅", "en": "Yoyogi-Uehara Station" },
-  { "ja": "中目黒駅", "en": "Nakameguro Station" },
-  { "ja": "大崎駅", "en": "Osaki Station" },
-  { "ja": "五反田駅", "en": "Gotanda Station" },
-  { "ja": "目黒駅", "en": "Meguro Station" },
-  { "ja": "恵比寿駅", "en": "Ebisu Station" },
-  { "ja": "渋谷駅", "en": "Shibuya Station" },
-  { "ja": "原宿駅", "en": "Harajuku Station" },
-  { "ja": "代々木駅", "en": "Yoyogi Station" },
-  { "ja": "新宿駅", "en": "Shinjuku Station" },
-  { "ja": "新大久保駅", "en": "Shin-Okubo Station" },
-  { "ja": "高田馬場駅", "en": "Takadanobaba Station" },
-  { "ja": "目白駅", "en": "Mejiro Station" },
-  { "ja": "池袋駅", "en": "Ikebukuro Station" },
-  { "ja": "大塚駅", "en": "Otsuka Station" },
-  { "ja": "巣鴨駅", "en": "Sugamo Station" },
-  { "ja": "駒込駅", "en": "Komagome Station" },
-  { "ja": "田端駅", "en": "Tabata Station" },
-  { "ja": "西日暮里駅", "en": "Nishi-Nippori Station" },
-  { "ja": "日暮里駅", "en": "Nippori Station" },
-  { "ja": "鶯谷駅", "en": "Uguisudani Station" },
-  { "ja": "上野駅", "en": "Ueno Station" },
-  { "ja": "御徒町駅", "en": "Okachimachi Station" },
-  { "ja": "秋葉原駅", "en": "Akihabara Station" },
-  { "ja": "神田駅", "en": "Kanda Station" },
-  { "ja": "有楽町駅", "en": "Yurakucho Station" },
-  { "ja": "浜松町駅", "en": "Hamamatsucho Station" },
-  { "ja": "田町駅", "en": "Tamachi Station" },
-  { "ja": "高輪ゲートウェイ駅", "en": "Takanawa Gateway Station" },
-  { "ja": "矢野口駅", "en": "Yanokuchi Station" },
-  { "ja": "稲城長沼駅", "en": "Inagi-Naganuma Station" },
-  { "ja": "南多摩駅", "en": "Minami-Tama Station" },
-  { "ja": "府中本町駅", "en": "Fuchu-Hommachi Station" },
-  { "ja": "分倍河原駅", "en": "Bubaigawara Station" },
-  { "ja": "西府駅", "en": "Nishi-Fuchu Station" },
-  { "ja": "谷保駅", "en": "Yaho Station" },
-  { "ja": "矢川駅", "en": "Yagawa Station" },
-  { "ja": "西国立駅", "en": "Nishi-Kunitachi Station" },
-  { "ja": "立川駅", "en": "Tachikawa Station" },
-  { "ja": "北府中駅", "en": "Kita-Fuchu Station" },
-  { "ja": "西国分寺駅", "en": "Nishi-Kokubunji Station" },
-  { "ja": "新小平駅", "en": "Shin-Kodaira Station" },
-  { "ja": "新秋津駅", "en": "Shin-Akitsu Station" },
-  { "ja": "成瀬駅", "en": "Naruse Station" },
-  { "ja": "町田駅", "en": "Machida Station" },
-  { "ja": "相原駅", "en": "Aihara Station" },
-  { "ja": "八王子みなみ野駅", "en": "Hachioji-Minami Station" },
-  { "ja": "片倉駅", "en": "Katakura Station" },
-  { "ja": "八王子駅", "en": "Hachioji Station" },
-  { "ja": "西大井駅", "en": "Nishi-Oi Station" },
-  { "ja": "四ツ谷駅", "en": "Yotsuya Station" },
-  { "ja": "吉祥寺駅", "en": "Kichijoji Station" },
-  { "ja": "三鷹駅", "en": "Mitaka Station" },
-  { "ja": "国分寺駅", "en": "Kokubunji Station" },
-  { "ja": "日野駅", "en": "Hino Station" },
-  { "ja": "豊田駅", "en": "Toyoda Station" },
-  { "ja": "西八王子駅", "en": "Nishi-Hachioji Station" },
-  { "ja": "高尾駅", "en": "Takao Station" },
-  { "ja": "御茶ノ水駅", "en": "Ochanomizu Station" },
-  { "ja": "中野駅", "en": "Nakano Station" },
-  { "ja": "高円寺駅", "en": "Koenji Station" },
-  { "ja": "阿佐ケ谷駅", "en": "Asagaya Station" },
-  { "ja": "荻窪駅", "en": "Ogikubo Station" },
-  { "ja": "西荻窪駅", "en": "Nishi-Ogikubo Station" },
-  { "ja": "武蔵境駅", "en": "Musashi-Sakai Station" },
-  { "ja": "東小金井駅", "en": "Higashi-Koganei Station" },
-  { "ja": "武蔵小金井駅", "en": "Musashi-Koganei Station" },
-  { "ja": "国立駅", "en": "Kunitachi Station" },
-  { "ja": "東中野駅", "en": "Higashi-Nakano Station" },
-  { "ja": "大久保駅", "en": "Okubo Station" },
-  { "ja": "千駄ケ谷駅", "en": "Sendagaya Station" },
-  { "ja": "信濃町駅", "en": "Shinanomachi Station" },
-  { "ja": "市ケ谷駅", "en": "Ichigaya Station" },
-  { "ja": "飯田橋駅", "en": "Iidabashi Station" },
-  { "ja": "水道橋駅", "en": "Suidobashi Station" },
-  { "ja": "浅草橋駅", "en": "Asakusabashi Station" },
-  { "ja": "両国駅", "en": "Ryogoku Station" },
-  { "ja": "錦糸町駅", "en": "Kinshicho Station" },
-  { "ja": "亀戸駅", "en": "Kameido Station" },
-  { "ja": "平井駅", "en": "Hirai Station" },
-  { "ja": "新小岩駅", "en": "Shin-Koiwa Station" },
-  { "ja": "小岩駅", "en": "Koiwa Station" },
-  { "ja": "新日本橋駅", "en": "Shin-Nihombashi Station" },
-  { "ja": "馬喰町駅", "en": "Bakurocho Station" },
-  { "ja": "西立川駅", "en": "Nishi-Tachikawa Station" },
-  { "ja": "東中神駅", "en": "Higashi-Nakagami Station" },
-  { "ja": "中神駅", "en": "Nakagami Station" },
-  { "ja": "昭島駅", "en": "Akishima Station" },
-  { "ja": "拝島駅", "en": "Haijima Station" },
-  { "ja": "牛浜駅", "en": "Ushihama Station" },
-  { "ja": "福生駅", "en": "Fussa Station" },
-  { "ja": "羽村駅", "en": "Hamura Station" },
-  { "ja": "小作駅", "en": "Ozaku Station" },
-  { "ja": "河辺駅", "en": "Kabe Station" },
-  { "ja": "東青梅駅", "en": "Higashi-Ome Station" },
-  { "ja": "青梅駅", "en": "Ome Station" },
-  { "ja": "宮ノ平駅", "en": "Miyanohira Station" },
-  { "ja": "日向和田駅", "en": "Hinatawada Station" },
-  { "ja": "石神前駅", "en": "Ishigami Station" },
-  { "ja": "二俣尾駅", "en": "Futamatao Station" },
-  { "ja": "軍畑駅", "en": "Gunji Station" },
-  { "ja": "沢井駅", "en": "Sawai Station" },
-  { "ja": "御嶽駅", "en": "Mitake Station" },
-  { "ja": "川井駅", "en": "Kawai Station" },
-  { "ja": "古里駅", "en": "Furusato Station" },
-  { "ja": "鳩ノ巣駅", "en": "Hatono Station" },
-  { "ja": "白丸駅", "en": "Shiromaru Station" },
-  { "ja": "奥多摩駅", "en": "Okutama Station" },
-  { "ja": "熊川駅", "en": "Kumagawa Station" },
-  { "ja": "東秋留駅", "en": "Higashi-Akiru Station" },
-  { "ja": "秋川駅", "en": "Akikawa Station" },
-  { "ja": "武蔵引田駅", "en": "Musashi-Hikida Station" },
-  { "ja": "武蔵増戸駅", "en": "Musashi-Masuko Station" },
-  { "ja": "武蔵五日市駅", "en": "Musashi-Itsukaichi Station" },
-  { "ja": "北八王子駅", "en": "Kita-Hachioji Station" },
-  { "ja": "小宮駅", "en": "Komiya Station" },
-  { "ja": "東福生駅", "en": "Higashi-Fussa Station" },
-  { "ja": "箱根ケ崎駅", "en": "Hakonegasaki Station" },
-  { "ja": "尾久駅", "en": "Oku Station" },
-  { "ja": "赤羽駅", "en": "Akabane Station" },
-  { "ja": "三河島駅", "en": "Mikawashima Station" },
-  { "ja": "南千住駅", "en": "Minami-Senju Station" },
-  { "ja": "北千住駅", "en": "Kita-Senju Station" },
-  { "ja": "綾瀬駅", "en": "Ayase Station" },
-  { "ja": "亀有駅", "en": "Kameari Station" },
-  { "ja": "金町駅", "en": "Kanamachi Station" },
-  { "ja": "板橋駅", "en": "Itabashi Station" },
-  { "ja": "十条駅", "en": "Jujo Station" },
-  { "ja": "北赤羽駅", "en": "Kita-Akabane Station" },
-  { "ja": "浮間舟渡駅", "en": "Ukima-Funado Station" },
-  { "ja": "八丁堀駅", "en": "Hatchobori Station" },
-  { "ja": "越中島駅", "en": "Etchujima Station" },
-  { "ja": "潮見駅", "en": "Shiomi Station" },
-  { "ja": "新木場駅", "en": "Shinkiba Station" },
-  { "ja": "葛西臨海公園駅", "en": "Kasai-Rinkai Park Station" },
-  { "ja": "東十条駅", "en": "Higashi-Jujo Station" },
-  { "ja": "王子駅", "en": "Oji Station" },
-  { "ja": "上中里駅", "en": "Kami-Nakazato Station" },
-  { "ja": "大井町駅", "en": "Oimachi Station" },
-  { "ja": "大森駅", "en": "Omori Station" },
-  { "ja": "蒲田駅", "en": "Kamata Station" }
-]
