--- src/styles/NextImage.scss
@@ -1,13 +1,0 @@
-.container {
-  position: relative;
-  width: 100%;
-  height: 100%;
-}
-
-.image {
-  object-fit: cover;
-}
-
-.top-rounded {
-  border-radius: 20px 20px 0 0;
-}
