--- src/styles/cafe-list.scss
@@ -1,25 +1,0 @@
-.row__cards {
-  overflow-x: scroll;
-  overflow-y: hidden;
-  // モニター
-  // width: 130em;
-  height: 20em;
-  // PC
-  @media screen and (max-width: 2200px) {
-    // width: 115em;
-  }
-  // mobile
-  @media screen and (max-width: 1200px) {
-    height: 17em;
-  }
-}
-
-.row__cards::-webkit-scrollbar {
-  display: none;
-}
-
-.row__picture {
-  object-fit: contain;
-  border-radius: 20px 20px 0% 0;
-  transition: transform 450ms;
-}
